added openai, tag, project routes

+++ controllers/projects_controller.py
@@ -25,8 +25,10 @@ cloudinary.config(
 projects = Blueprint("projects", __name__, url_prefix="/projects")
 
 
+
 @projects.route('/', methods=['GET'])
 def get_all_projects():
+    '''Get a list of all projects'''
     project_list = Project.query.all() # User.all() did not work (flask-marshmallow docs are wrong)
 
     return jsonify(projects_schema.dump(project_list)), 200 # OK
@@ -34,6 +36,7 @@ def get_all_projects():
 
 @projects.route('/<int:project_id>', methods=['GET'])
 def get_project_by_id(project_id: int):
+    '''Get a project by its ID number'''
     # project = Project.query.filter_by(id=project_id).first() # alternative way
     project = Project.query.get(project_id) # shorthand for querying by primary_key
     
@@ -43,17 +46,18 @@ def get_project_by_id(project_id: int):
     return jsonify(project_schema.dump(project)), 200 # OK
 
 
-# @jwt_required() # projectary led to error without the parentheses
 @projects.route('/', methods=['POST'])
+@jwt_required() # planectary led to error without the parentheses
 def create_project():
+    '''Create a new project'''
     project_fields = request.json
     # project_fields = project_schema.load(request.json)
     
     # print(request.json)
     # print(project_fields)
     # extract id of user from the JWT token
-    # request_id = get_jwt_identity()
-    user_id = 3
+    user_id = get_jwt_identity()
+    # user_id = 3
     
     title = project_fields.get('title')
     description = project_fields.get('description')
@@ -61,14 +65,19 @@ def create_project():
     demo_url = project_fields.get('demoURL')
     image_url = project_fields.get('imageURL')
     
-    print(project_fields.get('tags'))
+    print('Tags: ', project_fields.get('tags'))
     tags = project_fields.get('tags')
 
     try:
         image = project_fields.get('image')
+        
+        
         if image:
             response = cloudinary.uploader.upload(image, folder='projectshare')
+            print('Image: yes')
             image_url = response.get('url')
+        else:
+            print('Image: no')
 
 
         new_project = Project(
@@ -98,6 +107,7 @@ def create_project():
 @projects.route('/<int:project_id>', methods=['PUT'])
 # @jwt_required()
 def update_project(project_id: int):
+    '''Update a project by its ID'''
     # project_fields = project_schema.load(request.json)
     project_fields = request.json
 
@@ -113,6 +123,8 @@ def update_project(project_id: int):
     
     if not project:
         return abort(404, message="Project not found")
+    
+    
 
     project.title = project_fields.get('title')
     project.description = project_fields.get('description')
@@ -126,22 +138,43 @@ def update_project(project_id: int):
 
 
 @projects.route('/<int:project_id>', methods=['DELETE'])
-# @jwt_required()
+@jwt_required()
 def delete_project(project_id: int):
-    # user_id = get_jwt_identity()
+    '''Delete a project by its ID'''
+    user_id = get_jwt_identity()
+    print(user_id)
     # user_id = 3
     
     # # #Find it in the db
-    # user = User.query.get(user_id)
-    # if not user:
-    #     return abort(401, description="Invalid user")
+    user = User.query.get(user_id)
+    if not user:
+        return abort(401, description="Invalid user")
+    
     
     
     project = Project.query.get(project_id)
+    
+    print('project username and id:', project.user.username, project.user_id)
+    
     if not project:
-        return jsonify(message="That project does not exist"), 404
+        return jsonify(message="Project does not exist"), 404 # Not Found
+    
+    if user_id != project.user_id:
+        return jsonify(message="Unauthorized user"), 401 # Unauthorized (abort send backs html)
     
     db.session.delete(project)
     db.session.commit()
     
-    return jsonify(message="You deleted a project", id=project_id), 202 # Accepted
+    return jsonify(message="Project deleted", id=project_id), 202 # Accepted
+
+
+
+@projects.route('/tag', methods=['GET'])
+def get_projects_by_tag_name():
+    '''Get all projects that contain a specific tag in their tags list'''
+    tag_name = request.args.get('tag')
+
+    # projects = Project.query.filter(Project.tags.any(tag_name)).all()
+    projects = Project.query.join(Project.tags).filter(Tag.name == tag_name).all()
+    
+    return jsonify(projects_schema.dump(projects))
